+++ ruby/lib/essence/components/badge.rb
@@ -1,18 +1,30 @@
 # frozen_string_literal: true
 
-class Essence::Avatar < Essence::Essence
-  BASE = "border border-transparent rounded-full bg-gray-100 aspect-square inline-flex items-center justify-center font-medium text-gray-700"
+class Essence::Badge < Essence::Essence
+  BASE = "inline-flex items-center justify-center w-fit rounded-full border border-transparent font-medium transition duration-150 hover:opacity-90"
   SIZES = {
-    sm: "size-6 text-xs",
-    md: "size-8 text-sm",
-    lg: "size-12 text-base"
+    none: "",
+    sm: "text-[0.65rem] px-2 py-0.5 gap-1 min-w-8",
+    md: "text-xs px-2.5 py-1 gap-2 min-w-12",
+    lg: "text-sm px-4 py-1 gap-2 min-w-16"
+  }
+  KINDS = {
+    primary: "text-gray-900 border-gray-200",
+    success: "text-white bg-emerald-500",
+    critical: "text-white bg-rose-500",
+    warning: "text-white bg-amber-500",
+    info: "text-white bg-blue-500",
+    dark: "text-white bg-gray-900",
+    white: "text-gray-900 bg-white"
   }
 
-  attr_reader :attributes
   attr_reader :size
+  attr_reader :kind
+  attr_reader :attributes
 
-  def initialize(size: :md, **attributes)
+  def initialize(size: :md, kind: :primary, **attributes)
     @size = size
+    @kind = kind
     @attributes = attributes
     @attributes[:class] = construct_classes(@attributes[:class])
   end
@@ -21,11 +33,9 @@ class Essence::Avatar < Essence::Essence
     div(**attributes, &)
   end
 
-  def fallback(**attrs, &)
-    span(**attrs, &)
-  end
+  private
 
   def construct_classes(classes)
-    TAILWIND_MERGER.merge([ BASE, SIZES[size], classes ].compact)
+    TAILWIND_MERGER.merge([ BASE, SIZES[size], KINDS[kind], classes ].compact)
   end
 end