+++ ruby/CHANGELOG.md
@@ -4,7 +4,11 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 
 ### Unreleased
 
-- fix: Fixing `Button` component. It was mixed up with badge for some reason.
+### 0.2.2
+
+- fix: Fixing `Avatar` component. It was mixed up with badge for some reason. Oops!
+- fix: Fixing `Badge` component. It was mixed up with badge for some reason. Oops!
+- fix: Fixing `Button` component. It was mixed up with badge for some reason. Oops!
 
 ### 0.2.1